Update patch set 2
Patch Set 2:

>In VP8 multithreading encoder/decoder, the process of a macroblock can 
>only be started after its above macroblock is done, which guarantees 
>the correct processing order. However, TSan couldn't understand that.
Can you please point to the code that guaranteed the processing order previously?
In general TSan understands the happens-before relation between thread creation/execution/joining and acquisition and release of locks and atomic variables. TSan doesn't understand memory barriers, but I didn't notice those in the code, so I suspect the code was relying on an implicit assumption it's working on the architecture with a strong memory model (i.e. x86)

Patch-set: 2