Update patch set 2
Patch Set 2:

(1 comment)

Here is the place that guarantees the processing order.
In sync_read():
mb_col > (protected_read(mutex, last_row_current_mb_col) - nsync
The above condition guanrantees the current macroblock has to wait for the microblocks to finish in the above row.

Here is the place that I think TSan couldn't understand:
In thread_encoding_proc():
 if (mb_row == cm->mb_rows - 1)
    {
        sem_post(&cpi->h_event_end_encoding); /* signal frame encoding end */
    }

We only ask the thread that handles the last row to post the end of a frame encoding since we know the rows above the last one are all done at this moment.

Patch-set: 2