Update patch set 3
Patch Set 3:

(1 comment)

> Nico Weber from chrome/clang helpfully pointed out a related
 > presentation and bug [1][2]. The recommendation is to use
 > /FIintrin.h for now. I hadn't thought of that, but it's simpler, so
 > I'd prefer not to take this change since it can be worked around on
 > the compiler side.
 > 
 > [1] https://docs.google.com/presentation/d/1oxNHaVjA9Gn_rTzX6HIpJHP7nXRua_0URXxxJ3oYRq0/edit?usp=sharing
 > [2] http://llvm.org/PR19898

I'm aware of that LLVM bug, the problem is that right now the intrin.h header which comes with MSVC is very different than the one that clang uses, and force-including it everywhere may create random incompatibilities in other unrelated things.  The "proper" solution would be to also recognize #pragma intrinsic and just use the header that microsoft ships, but that doesn't seem like low hanging fruit.

At any rate, since we need this to build for now, I have committed the patch to Firefox's copy of libvpx.  I was hoping to get it upstreamed mostly for the benefit of other libvpx consumers, but if you decide to not take it, I guess we'll maintain the downstream patch for now.

Patch-set: 3
Reviewer: Gerrit User 1106279 <1106279@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0