Update patch set 2
Patch Set 2:

Thank you for looking at the code. I think there may be some misunderstanding here. I can clarify it.

vp8_encode_frame() is called by the main thread, and thread_encoding_proc() is called by created child threads. The main thread does row encoding just like the child threads.
if(mb_row == cm->mb_rows - 1)
  {
    sem_post(&cpi->h_event_end_encoding); /* signal frame encoding end */
  }
is placed in 2 places, but the sem_post only happens once - the thread that processes the last row does sem_post that is ensured by the condition: if(mb_row == cm->mb_rows - 1), and it can be the main thread or one of the child threads.

The situation here is that we don't want to add extra semaphores for each thread that is unnecessary at all, but I can see TSan warnings shown because of this implementation.

Patch-set: 2