vp8-denoiser: Add another level to increase_denoising.
Add another level to increase_denoising = (0, 1, 2).
Only difference between level 0 and 1 is the sum_diff_threshold.
New level (1, 2) corresponds to old level (0, 1).

Default level (level=1/medium) is kept the the same, and same
logic is kept for switching to level=2/high.
But for blocks detected as skin, we set it to the new level=0 (low).

So only change in behavior is for "skin" blocks. Now we allow
for weak/low denoising (level=0) for skin blocks, instead no-denoising
if skin block has motion.

Also the settings for SUM_DIFF_THRESHOLD and SUM_DIFF_THRESHOLD_HIGH
are modified/lowered.

Change-Id: Idfd2424c86f599af9a9ea6bc21c9e2b07a51eb4d