Update patch set 1
Patch Set 1:

(3 comments)

I believe Debargha is running tests to check the coding statistics for this.

It will definitely have an impact however. I highlighted the code where the partition contexts are restored, which happens after the context lookups that I have removed. I noticed this because I traced the rate costing of partition choices at rdopt vs packing the bistream, and they were off for everything apart from PARTITION_NONE blocks, which don't recursively call rd_pick_partition. With this change in, they are all in line with the bitstream packing.

Patch-set: 1