Update patch set 3
Patch Set 3:

> The basis of the problem is that tran_[low|high]_t depends on what
 > configuration the code is built in:
 > https://chromium.googlesource.com/webm/libvpx/+/master/vpx_dsp/vpx_dsp_common.h#26
 > 
 > This means that the intermediary buffers may be different sizes. In
 > this change:
 > https://chromium-review.googlesource.com/324310
 > the buffer-handling code is ifdef'd to account for the different
 > possibilities.
 > 
 > It may be the case that this code does not appear to be affected:
 > perhaps the buffers are only passed between optimized versions of
 > the function and through this happy accident the data is correctly
 > handled.
 > 
 > This would need to be fixed/verified by applying a fix similar to
 > the one above where the different size buffers are accounted for or
 > by finding a way to use different intermediate buffers for the 8
 > bit case.

I don't think the difference in 

 > The basis of the problem is that tran_[low|high]_t depends on what
 > configuration the code is built in:
 > https://chromium.googlesource.com/webm/libvpx/+/master/vpx_dsp/vpx_dsp_common.h#26
 > 
 > This means that the intermediary buffers may be different sizes. In
 > this change:
 > https://chromium-review.googlesource.com/324310
 > the buffer-handling code is ifdef'd to account for the different
 > possibilities.
 > 
 > It may be the case that this code does not appear to be affected:
 > perhaps the buffers are only passed between optimized versions of
 > the function and through this happy accident the data is correctly
 > handled.
 > 
 > This would need to be fixed/verified by applying a fix similar to
 > the one above where the different size buffers are accounted for or
 > by finding a way to use different intermediate buffers for the 8
 > bit case.

The size difference for tran_*_t doesn't seem to be an issue for these particular functions, as all tests still pass. However, the difference in tran_*_t is causing problems for the next CL I'm working on where I'm trying to make it possible to compile the decoder with highbit support but not the encoder. Currently, we would very much like to have highbit decoding support in chrome, but we do not need highbit encoding support. Currently, compiling the decoder with hithbit support uses up ~100kb, while compiling the encoder with highbit support uses up ~400kb. 400kb dead code is not acceptable for chrome.

As for making *this* CL acceptable for checking in, is there a method for asserting/verifying that no assumptions are violated?

Patch-set: 3
Reviewer: Gerrit User 1115926 <1115926@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0