Update patch set 1
Patch Set 1:

this is the only real "complicated" change. previously there was only one place where PREFIX was used so i replaced the whole stanza with the different language.

x86inc.asm added a use of PREFIX to cextern_naked. instead of copying the logic down there, i opted to define/undef (probably not necessary to undef) PREFIX based on the old logic, and leave the x86inc use of PREFIX

Patch-set: 1