Update patch set 2
Patch Set 2:

> Hi James, what do you think about using *.inc file for some common
 > code in the test?

This isn't the intent of .inc. That's meant for code that isn't meant to be built on its own, but #include'd. A base class should just live in a common source file and have a header to reference it.
The example given notes the few places where .inc makes sense:
"""
Examples are files that need to be included multiple times or platform-specific extensions that essentially are part of other headers.
"""

An example for the first half would be if some complex macros were used as a template for function names for instance.

Patch-set: 2