Do sub-pixel motion search in up-sampled reference frames
This patch was originally checked into nextgenv2 branch, here it is moved
to master branch.

Up-sampled the reference frames to 8 times in each dimension using the
8-tap interpolation filter. In sub-pixel motion search, use the up-sampled
reference frames to find the best matching blocks to increase the motion
search precision. This is enabled as a speed feature for speed 0 and
speed 1, and this is encoder-only improvement.

Master branch borg test results:
1. On derflr set,
Overall PSNR gain: 1.141%, and SSIM gain: 1.359%.
2. On stdhd set,
Overall PSNR gain: 0.669%, and SSIM gain: 0.771%.

Speed loss on several derf and HD clips was ~4.5%.

Change-Id: I7e66809aa76210fc32e0db15f5ad37f854093e37