vp9: Fix some logic for gf update in cyclic refresh.
Fix the logic for refreshing golden frame based on high motion.
The current comparison was incorrect and prevented this
(force update of gf on high motion) from ever being used.
For now keep this logic under a flag (and off for now) so as to
not change behavior, until further testing.

Change-Id: Ib5f0082159a428b0603b9534e4bcb6f83e4ccb25