Make superblock size variable at the frame level.
The uncompressed frame header contains a bit to signal whether the
frame is encoded using 64x64 or 128x128 superblocks. This can vary
between any 2 frames.

vpxenc gained the --sb-size={64,128,dynamic} option, which allows the
configuration of the superblock size used (default is 64). 64/128
will force the encoder to always use the specified superblock size.
Dynamic would enable the encoder to choose the sb size for each
frame, but this is not implemented yet (dynamic does the same as 128
for now).

Tile dimensions in the old tile coding scheme (without
--enable-ext-tile) are computed assuming 128x128 superblocks, even
if the encoder will code everything with 64x64 superblocks. This means
that a tile now need to be multiples of 128x128 pixels in size (as
opposed to 64x64 beforehand).

Tile boundaries with --enable-ext-tile are computed respecting the
superbock size used for the frame. The --tile-columns and --tile-rows
(which in turn really define tile witdh and tile height) are in
units of 64 pixels if --sb-size=64, or in units of of 128 if
--sb-size={128,dynamic}. If --sb-size={128,dynamic}, then the maximum
value of --tile-columns and --tile-rows (width/height) is reduced to
32.

Change-Id: Idc9beee1ad12ff1634e83671985d14c680f9179a