Some fixes/speed-ups on inter-intra part of ext-inter
Fixes an issue with rectangular inter-intra blocks
Searches only the best intra mode so far for the inter-intra search.
Includes various other refactoring and cleanups to enable fast mixing
of inter and intra predictors.

This is part one of overhauling on the ext-inter experiment. To be
continued in subsequent patches.

Change-Id: Id10ee100c78c6e00009a3a4f930a4435ef403a95