Update patch set 1
Patch Set 1:

I didn't really investigated this function or the c-code, so there's a chance that's at fault on its own or as well. With this change the results are at least consistent across machines and I have seen a case where an AVX machine has produced a bad block, though I couldn't directly link it to this function.

Patch-set: 1
Label: Code-Review=0