vp9: Fix an issue with setting initial_width/height.
Issue is that initial_width/height is not set on
codec initialization, so if an application changes the scale
on the first frame (before encoding, but after encoder-init),
and then changes the scale back up on a later frame, the assert
in vp9_set_internal_size() will trigger:
assert(cm->width <= cpi->initial_width)

Fix is to set initial_width upon codec initialization.

Modified TestInternalResizeChangeConfig to trigger the assert.

Change-Id: I0556c3b90e79f3af176da3d27972df748a10fcf9