vp9: Fix an issue with setting initial_width/height.
Issue is that initial_width/height is not set on
codec initialization, so if an application changes the scale
on the first frame (before encoding, but after encoder-init),
and changes the scale back up on a later frame, then the assert
(assert(cm->width <= cpi->initial_width) will be triggered.

Fix is to set initial_width in update_frame_size,
which is called on any change_config.

Modified TestInternalResizeChangeConfig to trigger the assert.

Change-Id: I0556c3b90e79f3af176da3d27972df748a10fcf9