Update patch set 1
Patch Set 1:

Thanks Julia. Another way is in handle_inter_mode(), adding the cost of an obmc=0 flag back to the rd cost of interinter and interintra modes. 

But this needs more efforts because we have to make the logic of RDO and bitstream syntax consistent, I can take care of it. In the bitstream, the obmc flag is coded first, if it is 0 then we continue to check the interinter and interintra modes. But in the RDO, both the rate calculation and the logic are a bit messy: (1) the rate of obmc=0 is omitted if allow_obmc = 0; (2) interinter mode is first examined, only if it is not better than inter mode, obmc is checked, otherwise set allow_obmc as 0; (3) for interintra mode, allow_obmc is always 0. Therefore if allow_obmc is set to 0 due to interintra/interinter, we need to add the rate of obmc flag.

Patch-set: 1