Update patch set 1
Patch Set 1:

My problem with this is not the specifics but if it is the right way to go.

CQ inherently behaves like VBR on harder content anyway so the only clips affected by the cq level are the easy sections with many smooth blocks. Hence this just makes CQ lower by the back door. The real problem surely is the value being set too high in the first place.

An alternate approach to the cq mode to save bits in easier clips / chunks may be to assign each clip or chunk or kf group to a complexity / perceptual bin based on first pass stats and then apply an alteration to the target rate based on the bin. So easy sections might be encoded at say 25% of the nominal bandwidth and the hardest 150% or some such. Thereafter though the code would act as vbr and try and hit the given (modified) rate for the chunck / section.

Patch-set: 1