Fix segfault with --cpu-used >= 3 and ext-refs.
With ext-ref enabled, it is possible that when trying to encode the
first true altref frame after a keyframe, the previous altref frame
(alias for the keyframe) is the same as one of the new LAST{2,3,4}
reference frames, and hence cpi->ref_frame_flags will have the altref
bit clear, as computed by get_ref_frame_flags in encoder.c.

sf->alt_ref_search_fp forces the previous altref frame to
be used as the only possible  reference when encoding a new altref
frame, but due to cpi->ref_frame_flags, some buffers will not be
initialized (see rdopt.c:7689 yv12_mb), leading to a segfault.

Given that this is only a speed feature, the fix I added is simply to
ignore sf->alt_ref_search_fp if the altref buffer is not initialized
due to cpi->ref_frame_flags.

Change-Id: I4da1991667614009da5d3061a6316c0d5dbc6c0c