Update patch set 1
Patch Set 1:

Sorry for the confusion on the review.

If the issue is as what has been addressed in the comments, I would suggest to change the settings in

static int get_ref_frame_flags(const VP10_COMP *cpi)

in encoder.c, in a way such that the altref bit keeps set but to remove the bit set for LAST{2,3,4}, whichever is the same as altref, hence the bit setting for altref in cpi->ref_frame_flags would be able to then set in the same way as when ext-ref is not enabled.

Patch-set: 1
Label: Code-Review=0