Update patch set 2
Patch Set 2:

The test matches up with the _c but fails for sse2 and ssse3. The sse2 passes when it runs the same code for the first and second loop. I added a RandI9 function to generate 9 bit input for it, but it fails in the same way when using Rand8. Are the constraints OK? Is the test overly aggressive?

Patch-set: 2
Label: Code-Review=0