Fixed 8K two pass encoder crash.
Bug found by Yunqing relating to the correction for size at 8K and above
in get_twopass_worst_quality().

The basis for the correction was changed to the linear size relative to 108P
as a baseline and the adjustment has been clamped to prevent problems
at extreme images sizes.

For 1080P the results on our test sets were neutral but the low res and mid
res sets saw a small gain (0.1%-0.2% average).

I would also expect gains on 4k and larger content where the previous
correction was overly aggressive.

Change-Id: I30b026b5f4535e9601e3178d738066459d19c8fb