Fix 8K crash bug.
Bug found by Yunqing where the correction for size at 8K and above
would force the error divisor to go -ve.

The basis for the correction was changed to the linear size relative to 108P
as a baseline and the adjustment has been clamped to prevent this sort of
problem even at extreme images sizes.

Change-Id: I30b026b5f4535e9601e3178d738066459d19c8fb