Update patch set 10
Patch Set 10:

This version drops the dependency on the inttypes.h removal since that's not a safe change to make. I've also removed the module map for now since it can't be enforced (and on closer inspection, the module map I specified was dependent on another hack I'd added to move the header files up one subdirectory, which I don't want to be pushing just yet).

Commit message updated. Everything seems to work pulling this version of the code as a pre-built .framework into my OGVKit wrapper library via CocoaPods...

Patch-set: 10
Label: Code-Review=0