Make the length of the backward ref group adjustable
This is for the bidir-pred experiment. Previously the length of the
backward ref group is fixed at 2, i.e. one bi-directionally predicted
frame can be inserted every other frame. With this patch, the length
becomes adjustable, i.e. any positive number may be specified, but
the use of the backward ref will be turned off if the backward ref
group is larger than the golden ref group.

Change-Id: I9ff7daefabcee2d79ab665b8710f6e7ad0bee0a3