VP9: always keep the last coded frame in frame buffers
In current VP9 implementation, if the last coded frame won't be used as a
reference for the next frame, its frame buffer might be re-assigned to the
next frame. However, this last frame may be needed later for other purpose,
and it should be kept. This patch made VP9 work in this situation and added
a parameter in VP9_COMMON to indicate which buffer to keep while finding a
free buffer for the next frame.

Change-Id: I9a83b01b52d4065aaea10e3ad5251b30145ef3e8