Update patch set 1
Patch Set 1:

> Note: I had a downstream user of my OGVKit library (which uses the
 > dynamic VPX.framework build) warn that their app submission was
 > automatically rejected by Apple because the VPX.framework's
 > Info.plist contained an invalid bundle version.
 > 
 > Unfortunately I didn't catch this earlier because XCode doesn't
 > care about this validity when you link and run programs -- just
 > when you submit the surrounding app to the store.
 > 
 > The bundle version is limited to three period-separated integers,
 > so the version extras with the git commit count since tag & commit
 > id can't simply be injected as fourth or fifth integers.
 > 

Well, that's silly. Oh well. What do other open source projects do with their frameworks? Can we have a custom property (GitVersion ?) that includes the entire version string? 

 > If using a bare "1.5.0" etc is not acceptable for libvpx upstream,
 > then I could devise a packed format, say:
 > 
 > 1.5.0-906-g656f9c4 -> 1005000.906.106363332
 > 
 > as (major * 1000000 + minor * 1000 + patch), (git offset from tag),
 > (first 7 hex digits of git hash converted to decimal)
 > 
 > Let me know if that's preferred and I'll update the patch.

No need to go this far. The motivation for having the entire git version is simply to make it easier to track down the exact repo hash used for a framework build. It makes it easier to determine what's broken, when it was broken, and when it's fixed etc etc.

I think a custom plist entry is the way to go, but I'm curious about other open source frameworks. Do they let this slide and just go with what's obtainable from the first three numbers?

Patch-set: 1
Label: Verified=0