Fix timebase for WebM files
CL:https://chromium-review.googlesource.com/#/c/322231 broke setting
the timebase for all files. But changing ivf to use rate and scale will
still be an issue for checking the psnr with FFmpeg. In the end I think
this is because of having two timing mechanisms in ivf files.

This change will fix timebase to work with webm files.

BUG=https://bugs.chromium.org/p/webm/issues/detail?id=1228

Change-Id: I53fb4baedc15e504d69b5df733214c32bfb7fa9a