Update patch set 2
Patch Set 2:

My comments on patch set 2 was not correct.
The comments are from C++ function overloading, e.g.

void f(int x);         // (1)
void f(const int x);   // (2)

are not allowed.  The "const" qualifier is ignored by compiler.  But in C, (1) or (2) can be used independently. Inside function (2), x is not allowed to change.  This is still useful.

Patch-set: 2