Update patch set 2
Patch Set 2:

(2 comments)

Using the uniform quantizer on average gives 0.2% better compression performance than the quant_b, which assumes an extended zero bin during the quantization process.

The difference really lies in the places where one gets a lot original coefficients that are close to the boundary between index (-1, 0) or (0, 1). This largely depends on the prediction quality and the target bit-rate. Hence typical observations are that uniform quantizer works better in several operating points of a given clip.

When the original coefficient is outside the extended zero bin, or it is within (-1/2, 1/2), the quant_b natural subsumes to the uniform quantizer. Hence the two have identical performance in such situations.

Patch-set: 2
Label: Code-Review=0