Disable iOS bitcode unless passed --enable-bitcode
Bitcode doesn't work for anything built from assembly, and recent
changes in Apple's systems seem to be rejecting app submissions now
both when the app has bitcode enabled (failure to link the dylib)
and when the app has bitcode disabled (for shared library that would
still have it in there).

Switching bitcode off unless an explicit --enable-bitcode is passed.

Change-Id: I29600dbe90c543960748561e41149170e6fb4c73