Add a hardware compatibility feature
There is a known issue in some hardware decoders:
in high bit depth settings, for chroma channel, when the upper-left
pixel is not available for intra prediction, its value is not
set as the correct default values.

To work around this issue, the added feature forces the encoder
to use intra modes that do not use the upper-left reference pixel
for chroma in high bit depth settings.

Change-Id: Ic15a938a3107e1b85e96cb7903a5c4220986b99d