vp9: Skip computation of y_sad_g for static blocks in choose partition.
Previously, y_sad_g is computed first. This patch changes the order to
compute y_sad first (using vp9_int_pro_motion_estimation which also
returns an estimated mv for the SB). When the estimated mv is (0,0), the
SB is kind of static and we can skip the computation of y_sad_g which
saves ~1% cycles for some clips with < 0.2% overall PSNR loss.

Change-Id: I781d7c7314ddbe1ce455ed0217b4f417e9af8208