Update patch set 5
Patch Set 5:

> I'm not sure anything really warranted this abstraction. the
 > functions in acm_random.h are simple to use and easy to understand
 > within the code.

The primary reason for this was the simple handling of filling multidimensional arrays with constrained random values which we use a lot and with AcmRandom it requires boilerplate that distracts from the test. The same goes for the Array wrapper in the related patch.

Contrast test/blend_mask6_test.cc to for example test/variance_test.cc, if you prefer the style in variance_test then ok I will live with that. It's just that I think tests should be simple to write so people are encouraged to write them, and it should be simple to achieve high coverage. I was hoping to work towards this with these patches.

Patch-set: 5
Reviewer: Gerrit User 1137274 <1137274@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0