Update patch set 2
Patch Set 2:

Linfeng and Scott have been working in some of this code recently. I'd like to verify the performance on some celeron machines.

Note we've made some decisions in the past to bias against big-core machines when dealing with ssse3 as it can have quite a bit range in performance. We're looking at trying to address both in certain cases with an additional cpuid check.

For reference take a look at:

https://chromium-review.googlesource.com/#/c/351273/

and the patch progression here:

https://chromium-review.googlesource.com/#/c/295244/

Patch-set: 2