Update patch set 2
Patch Set 2:

The original 1517fb7 merged and reverted from master, because of the failing of byte_alignment_test on Windows 64-bit. Patch set 1 is the code rebase to 1517fb7, and Patch 2 is the fix. In the fix, using 6+(X86_SUBPIX_VFILTER_PREFER_SLOW_CELERON*3) instead of 6+(ARCH_X86_64*3) for number of used regular registers, so that when X86_SUBPIX_VFILTER_PREFER_SLOW_CELERON is gone or set to 0, the correct number will be restored.

Patch-set: 2
Label: Code-Review=0