Remove sending extra info. for show_exsiting_frame
Originally we need to send the refresh flag and the virtual indices
mapping of the reference frame buffer for show_existing_frame to
resolve encoder-decoder mismatch.

To remove sending this infomation, we update the the virtual indices
of the reference frame buffer after the last_bipred_frame is encoded,
and therefore the decoder will receive the updated mapping at the next
non-existing frame.

Change-Id: I63d41ee6ea99884798f0778b789d2701e2f2d3e0